5 - 3 : convert string to double

+++ intro/05/Calculator/Calculator/Program.cs
@@ -8,10 +8,10 @@ namespace Calculator
         {
             Console.WriteLine("첫 번째 숫자를 입력하세요.");
             string userInput1 = Console.ReadLine();
-            int number1 = int.Parse(userInput1);
+            double number1 = double.Parse(userInput1);
             Console.WriteLine("두 번째 숫자를 입력하세요.");
             string userInput2 = Console.ReadLine();
-            int number2 = int.Parse(userInput2);
+            double number2 = double.Parse(userInput2);
 
             Console.Write(number1);
             Console.Write(" + ");
